Create FifteenTags.cpp
In the "Human" mode, the user interacts with the game by entering the value of the tag they want to move. The objective is to rearrange the numbers on the board in ascending order, leaving an empty space for sliding the tags. The game keeps track of the number of moves and displays the time taken to complete the puzzle.

In the "Computer" mode, the program automatically solves the puzzle using an algorithm based on the A* search algorithm. The algorithm calculates the Manhattan distance heuristic to estimate the distance of each tag from its target position. The computer makes moves until the puzzle is solved and displays the number of moves and time taken.